# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Ryno H Creations
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Painting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl: string | null;
  featured: boolean;
  createdAt: string;
}

export interface PaintingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  price: number;
  featured?: boolean;
}

export interface PaintingUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  price?: number;
  featured?: boolean;
}

export interface GalleryStats {
  total: number;
  minPrice: number;
  maxPrice: number;
  avgPrice: number;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  token: string;
}