# lib/api-zod/src/generated/types/galleryStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Ryno H Creations
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryStats {
  total: number;
  minPrice: number;
  maxPrice: number;
  avgPrice: number;
}