# lib/api-zod/src/generated/types/paintingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Ryno H Creations
 * OpenAPI spec version: 0.1.0
 */

export interface PaintingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  price: number;
  featured?: boolean;
}