# lib/db/src/schema/paintings.ts
import { pgTable, text, serial, timestamp, boolean, numeric } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const paintingsTable = pgTable("paintings", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  description: text("description"),
  price: numeric("price", { precision: 10, scale: 2 }).notNull(),
  imageUrl: text("image_url"),
  featured: boolean("featured").notNull().default(false),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertPaintingSchema = createInsertSchema(paintingsTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
});
export type InsertPainting = z.infer<typeof insertPaintingSchema>;
export type Painting = typeof paintingsTable.$inferSelect;